requirements.txt

--trusted-host thefirstjedi.asuscomm.com
--index-url http://thefirstjedi.asuscomm.com:3141/gymnos/stable/+simple
--extra-index-url http://thefirstjedi.asuscomm.com:3141/gymnos/staging/+simple

# Gymnos Packages
gymnosfirestoreapi==0.3

# PyPi Package
opencv-python==4.1.0.25
numpy==1.17.0
Keras==2.2.4
Pillow==6.1.0
matplotlib==3.1.1
